requirements.txt
flask
flask-cors
tensorflow
numpy
pandas
scikit-learn
mysql-connector-python
joblib
gunicorn